% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Flex.R
\name{Flex}
\alias{Flex}
\title{Flex}
\value{
A Flex object
}
\description{
Get flex reports using Intreractive Brokers Flex API.
}
\details{
The Flex class is used to get flex reports from Interactive Brokers.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{token}}{The token to be used for the IB Flex API}

\item{\code{query}}{The query to be used for the IB Flex API}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Flex-new}{\code{Flex$new()}}
\item \href{#method-Flex-send}{\code{Flex$send()}}
\item \href{#method-Flex-get}{\code{Flex$get()}}
\item \href{#method-Flex-get_flex_report}{\code{Flex$get_flex_report()}}
\item \href{#method-Flex-clone}{\code{Flex$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flex-new"></a>}}
\if{latex}{\out{\hypertarget{method-Flex-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Flex object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flex$new(token, query)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{token}}{The token to be used for the IB Flex API}

\item{\code{query}}{The query to be used for the IB Flex API}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The Flex object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flex-send"></a>}}
\if{latex}{\out{\hypertarget{method-Flex-send}{}}}
\subsection{Method \code{send()}}{
Send the request to the IB Flex API
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flex$send()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The new token to be used for the IB Flex API
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flex-get"></a>}}
\if{latex}{\out{\hypertarget{method-Flex-get}{}}}
\subsection{Method \code{get()}}{
Get the flex report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flex$get(newtoken)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newtoken}}{The new token to be used for the IB Flex API}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The flex report
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flex-get_flex_report"></a>}}
\if{latex}{\out{\hypertarget{method-Flex-get_flex_report}{}}}
\subsection{Method \code{get_flex_report()}}{
Get the flex report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flex$get_flex_report()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The flex report
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flex-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Flex-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flex$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
