% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{ibflexr-package}
\alias{ibflexr}
\alias{ibflexr-package}
\title{ibflexr: Managae flex reports and prepaare data for quant analysis.}
\description{
Dowbnload flex reports from Interactive Brokers and prepare data for quant analysis.
}
