% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Strategy.R
\name{Strategy}
\alias{Strategy}
\title{Strategy}
\value{
A Strategy object
}
\description{
Strategy class is used to extract and clean data from the flex reports.
}
\details{
The Strategy class is used to extract and clean data from the flex reports.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{flex_reports_xml}}{list of XML flex reports}

\item{\code{start_date}}{The start date of the strategy}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Strategy-new}{\code{Strategy$new()}}
\item \href{#method-Strategy-extract_node}{\code{Strategy$extract_node()}}
\item \href{#method-Strategy-summary_cfd_trades}{\code{Strategy$summary_cfd_trades()}}
\item \href{#method-Strategy-calculate_nav_units}{\code{Strategy$calculate_nav_units()}}
\item \href{#method-Strategy-clone}{\code{Strategy$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-new"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Strategy object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$new(flex_reports_xml, start_date = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{flex_reports_xml}}{list of XML flex reports}

\item{\code{start_date}}{The start date of the strategy}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The Strategy object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-extract_node"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-extract_node}{}}}
\subsection{Method \code{extract_node()}}{
Extract node from the XML report. The function clean data so that it can
be easly used for further analysis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$extract_node(node, filter_date = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{node}}{The node to be extracted}

\item{\code{filter_date}}{Filter date after start_date for column tradeDate if tradeDate column exist in data.table}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The extracted node in data.table format
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-summary_cfd_trades"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-summary_cfd_trades}{}}}
\subsection{Method \code{summary_cfd_trades()}}{
Summarize CFD trades
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$summary_cfd_trades()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The summarized CFD trades
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-calculate_nav_units"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-calculate_nav_units}{}}}
\subsection{Method \code{calculate_nav_units()}}{
Calculate NAV units
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$calculate_nav_units(
  benchmark_symbol = NULL,
  start_date = self$start_date,
  unit = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{benchmark_symbol}}{The benchmark symbol}

\item{\code{start_date}}{The start date}

\item{\code{unit}}{The unit}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The NAV units
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Strategy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Strategy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Strategy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
